Extracted `ByteBufferBackedInputFile` as a standalone class
