+++ hudi-common/src/main/java/org/apache/hudi/common/table/log/block/HoodieParquetDataBlock.java
@@ -168,40 +168,4 @@ public class HoodieParquetDataBlock extends HoodieDataBlock {
 
     return records;
   }
-
-  static class ByteBufferBackedInputFile implements InputFile {
-    private final byte[] buffer;
-    private final int offset;
-    private final int length;
-
-    ByteBufferBackedInputFile(byte[] buffer, int offset, int length) {
-      this.buffer = buffer;
-      this.offset = offset;
-      this.length = length;
-    }
-
-    ByteBufferBackedInputFile(byte[] buffer) {
-      this(buffer, 0, buffer.length);
-    }
-
-    @Override
-    public long getLength() {
-      return length;
-    }
-
-    @Override
-    public SeekableInputStream newStream() {
-      return new DelegatingSeekableInputStream(new ByteBufferBackedInputStream(buffer, offset, length)) {
-        @Override
-        public long getPos() {
-          return ((ByteBufferBackedInputStream) getStream()).getPosition();
-        }
-
-        @Override
-        public void seek(long newPos) {
-          ((ByteBufferBackedInputStream) getStream()).seek(newPos);
-        }
-      };
-    }
-  }
 }
